Compute row actions width instead of using ref (#31)
## Description
The ref strategy doesn't work when a table is in a tab, and causes the row actions to be unusably cut off. When the table isn't displayed, it gets a offsetWidth of 0, which becomes 50 (row min width) when tabbed to. We use the old strategy of computing the row actions width while preserving the snap to width behavior.

Fixes CAP-2328

## Test plan
`yarn storybook`

Before:
<img width="841" alt="Screenshot 2023-04-20 at 1 53 53 AM" src="https://user-images.githubusercontent.com/110051501/233314000-dea35696-69a3-402f-9c8e-ff2ba765af16.png">

After:
<img width="840" alt="Screenshot 2023-04-20 at 1 54 47 AM" src="https://user-images.githubusercontent.com/110051501/233314018-3adf9cd1-0c7a-4a31-aeca-acbc170d73a9.png">


Co-authored-by: None <48659329+aviator-app[bot]@users.noreply.github.com>